[enh] utils: implement a context manager for a SQLite cursor
usage in a with statement [2]::

    with SQLiteCursor('test.db') as cur:
        print(cur.execute('select sqlite_version();').fetchall()[0][0])

[1] https://docs.python.org/3/library/stdtypes.html#context-manager-types
[2] https://docs.python.org/3/reference/compound_stmts.html#with
